+++ OnlineHostelMgmt/src/main/java/com/cg/services/IRoomService.java
@@ -12,11 +12,11 @@ public interface IRoomService {
 	
 	public Room addRoom(RoomDTO roomdto) throws HostelNotFoundException;
 	
-	public List<Room> getRoomsByHostelId(Long hostel_id)throws HostelNotFoundException,RoomNotFoundException;
+	public List<Room> getRoomsByHostelId(Integer hostel_id)throws HostelNotFoundException,RoomNotFoundException;
 	
-	public List<Room> getRoomsByFloorAndHostelId(Integer floor,Long hostel_id) throws HostelNotFoundException,FloorNotFoundException,RoomNotFoundException ;
+	public List<Room> getRoomsByFloorAndHostelId(Integer floor,Integer hostel_id) throws HostelNotFoundException,FloorNotFoundException,RoomNotFoundException ;
 	
-	public List<Room> getRoomsAvailableByHostelId(Long hostel_id)throws HostelNotFoundException,RoomNotFoundException;
+	public List<Room> getRoomsAvailableByHostelId(Integer hostel_id)throws HostelNotFoundException,RoomNotFoundException;
 	
 	public List<Room> getRoomsAvailable() throws RoomNotFoundException;
