+++ OnlineHostelMgmt/src/test/java/com/cg/testHostel/TestViewAllHostel.java
@@ -38,9 +38,9 @@ public class TestViewAllHostel {
 	@BeforeEach
 	public void beforeEach(){
 		
-		list1.add(new Hostel((long)1, "NHSM Hostel", "8918273646", "girls", "kolkata, west bengal", 32000.0, null));
-		list1.add(new Hostel((long)2, "GNIT Hostel", "9898172634", "boys", "sodepur, west bengal", 30000.0, null));
-		list1.add(new Hostel((long)3, "Techno Hostel", "9991723361", "girls", "kolkata, west bengal", 42000.0, null));		
+		list1.add(new Hostel(1, "NHSM Hostel", "8918273646", "girls", "kolkata, west bengal", 32000.0, null));
+		list1.add(new Hostel(2, "GNIT Hostel", "9898172634", "boys", "sodepur, west bengal", 30000.0, null));
+		list1.add(new Hostel(3, "Techno Hostel", "9991723361", "girls", "kolkata, west bengal", 42000.0, null));		
 	}
 	
 	@Test
